requirements.txt
yfinance==0.2.66
numpy
pandas==2.3.3
pymongo==4.15.3
pytz==2025.1
PyYAML==6.0.3
scikit_learn==1.7.2
ta==0.11.0
python-json-logger
Flask==3.1.2
flask_pymongo==3.0.1
flask-cors==6.0.1
psutil==5.9.0
Flask-JWT-Extended==4.7.1
firebase_admin==7.1.0
